Make balances `u128` (#2498)
* WIP

* WIP: It compiles

* Migrate gas price to u128

* Fix priority fee math. Serialize gas price as string

* Update auto-generated files

* Update changelog

* Fix compilation of excluded crates

* WIP: Migrate all balances to Amount struct

* Minor fix in rewards

* Make from_primitive const, add Amount::MAX

* Some progress for macro

* One error!

* Keep going

* Keep spreading amount

* Compiles!

* Keep updating tests

* Making progress

* Making progress part 2

* Lint passes!

* Clean up and satisfy CI

* More tests are passing

* TRy this

* Fix openapi spec

* Updated chain hash

* Insta review

* Fmt and happy switcheroo

* Print key in error!

* More logging

* Fmt and useless conversion

* Remove useless conversion in Genesis

* Ignore max_fee tests

* Lint passes again after merge

* Revert switcheroo

* Use `Amount::MAX`

* Saturating gas_used delta

* Proper feature gating of sov-uniqueness

* Clean up trait

* More clean ups

* Use u64 from gas values

* Try to temp fix genesis configs fro demo data

* After merge

* Chain hash is updated

* Switcherooo, my friend!

* Update with the latest PR update in web3-js

* Fixing demo-rollup readme's, hopefully

* Decompose gas price update into limbs

* Use amount for bank and prover incentives genesis

* Incorporing latest gas charge patterns (#2502)

* Incorporing latest gas charge patterns

* PR comments

* Docs + testfix

* CHANGELOG

* nit

* PR comments

* Fix

* Remove test after bad merge

* Try fixes!

* Fix fmt after merge

* two more fixes for demo-rollup

* Another fix for mock-demo rollup

---------

+++ crates/full-node/sov-sequencer/src/preferred/mod.rs
@@ -272,7 +272,7 @@ where
         // We pass the node state root explicitly because retrieving it is
         // fallible, so it's convenient to front-load the error-checking.
         node_state_root: <S::Storage as Storage>::Root,
-        minimum_profit_per_tx: u64,
+        minimum_profit_per_tx: u128,
     ) {
         self.start_rollup_block_inner(
             visible_increase,
@@ -295,7 +295,7 @@ where
         visible_increase: VisibleSlotNumberIncrease,
         visible_slot_number_after_increase: Option<VisibleSlotNumber>,
         node_state_root: <S::Storage as Storage>::Root,
-        minimum_profit_per_tx: u64,
+        minimum_profit_per_tx: u128,
     ) {
         let InternalState::Idle {
             mut checkpoint,
@@ -933,7 +933,7 @@ pub struct PreferredSequencerConfig {
     /// The minimum fee that the preferred sequencer is willing to accept, denominated in rollup tokens. Defaults to zero.
     /// Sequencers should set this to a non-zero value if they wish to cover their DA costs.
     #[serde(default)]
-    pub minimum_profit_per_tx: u64,
+    pub minimum_profit_per_tx: u128,
     /// The size of the Tokio channel used to stream events.
     ///
     /// Don't deviate from the default unless you know what you're doing.
